# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "movie-reviews-api",
        "description": "API documentation movie-reviews-api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/medias": {
            "get": {
                "tags": [
                    "media"
                ],
                "summary": "get medias.",
                "operationId": "85663cb33f26f7c99e5bedf3e9a7595b",
                "parameters": [
                    {
                        "name": "search_term",
                        "in": "query",
                        "description": "search_term",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "number of items in a page",
                        "schema": {
                            "format": "int64",
                            "default": 10
                        }
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "filter score",
                        "schema": {
                            "format": "float",
                            "default": ""
                        }
                    },
                    {
                        "name": "likes_count",
                        "in": "query",
                        "description": "filter likes_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "dislikes_count",
                        "in": "query",
                        "description": "filter dislikes_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "reviews_count",
                        "in": "query",
                        "description": "filter reviews_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "release_date",
                        "in": "query",
                        "description": "filter release_date",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "filter category name",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort by",
                        "schema": {
                            "format": "string",
                            "default": "",
                            "enum": [
                                "newest",
                                "oldest",
                                "newest-release",
                                "oldest-release",
                                "most-likes",
                                "least-likes",
                                "most-dislikes",
                                "least-dislikes",
                                "most-reviews",
                                "least-reviews",
                                "best-reviewed",
                                "worst-reviewed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "media"
                ],
                "summary": "create new media.",
                "operationId": "1b067a94bf495b7adc019bbc63a6a4b8",
                "requestBody": {
                    "description": "request body",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of media",
                                        "default": ""
                                    },
                                    "url": {
                                        "description": "url of media",
                                        "default": ""
                                    },
                                    "release_date": {
                                        "description": "release_date of media",
                                        "default": ""
                                    },
                                    "category_name": {
                                        "description": "category_name of media",
                                        "default": ""
                                    },
                                    "summary": {
                                        "description": "summary of media",
                                        "default": ""
                                    },
                                    "storyline": {
                                        "description": "storyline of media",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "media created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}